feat: rules use a `Vec<String>` for values and project re-structured
