+++ src/rules/best_practises/code_complexity.rs
@@ -1,0 +1,6 @@
+// Mandatory
+const RULE_ID: &str = "code-complexity";
+const DEFAULT_SEVERITY: &str = "warn";
+
+// Specific
+const DEFAULT_COMPLEXITY: i32 = 7;
