+++ src/rules/best_practises/mod.rs
@@ -1,0 +1,1 @@
+pub mod code_complexity;
