+++ src/rules/mod.rs
@@ -1,0 +1,2 @@
+pub mod best_practises;
+pub mod rules_base;
